isolate code that collects env var info
